implement POST /api/v1/:shop_id/rate + test

+++ rvnx_doener_service/ent/event.go
@@ -16,7 +16,7 @@ import (
 type Event struct {
 	config `json:"-"`
 	// ID of the ent.
-	ID int `json:"id,omitempty"`
+	ID uint64 `json:"id,omitempty"`
 	// Created holds the value of the "created" field.
 	Created time.Time `json:"created,omitempty"`
 	// EventType holds the value of the "event_type" field.
@@ -58,7 +58,7 @@ func (e *Event) assignValues(columns []string, values []interface{}) error {
 			if !ok {
 				return fmt.Errorf("unexpected type %T for field id", value)
 			}
-			e.ID = int(value.Int64)
+			e.ID = uint64(value.Int64)
 		case event.FieldCreated:
 			if value, ok := values[i].(*sql.NullTime); !ok {
 				return fmt.Errorf("unexpected type %T for field created", values[i])
